refactor(swc/plugin_runner): split imported_fn into categories
