+++ crates/swc_plugin_runner/src/imported_fn/set_transform_result.rs
@@ -1,0 +1,11 @@
+use crate::{context::HostEnvironment, memory_interop::copy_bytes_into_host};
+
+/// Set plugin's transformed result into host's enviroment.
+/// This is an `imported` fn - when we instantiate plugin module, we inject this
+/// fn into pluging's export space. Once transform completes, plugin will call
+/// this to set its result back to host.
+pub fn set_transform_result(env: &HostEnvironment, bytes_ptr: i32, bytes_ptr_len: i32) {
+    if let Some(memory) = env.memory_ref() {
+        (*env.transform_result.lock()) = copy_bytes_into_host(memory, bytes_ptr, bytes_ptr_len);
+    }
+}
